+++ src/router.js
@@ -3,6 +3,7 @@ import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
 import Layout from './layout';
 import Main from './main';
+import Feed from './coverage/feed';
 import List from './coverage/list';
 import Coverage from './coverage/coverage';
 import File from './coverage/file';
@@ -16,6 +17,7 @@ export default (
         <Route path="/coverage/:source/:owner/:name" component={Coverage} />
         <Route path="/coverage/:source/:owner/:page?" component={List} />
         <Route path="/coverage/:page?" component={List} />
+        <Route exact path="/feed" component={Feed} />
         <Route exact path="/" component={Main} />
         <Route path="*" component={Error} />
       </Switch>